requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pandas==2.1.3
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.25.2
darts==0.25.0
prophet==1.1.4
pystan==2.19.1.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
langchain==0.0.350
streamlit==1.28.2
altair==5.2.0
plotly==5.17.0
pytest==7.4.3
httpx==0.25.2
bcrypt==4.1.2
python-multipart==0.0.6
google-generativeai==0.3.2